Continued GraphEdges and Began Testing

+++ src/ch/epfl/javelo/Q28_4.java
@@ -27,7 +27,7 @@ public final class Q28_4 {
      * @return le double correspondant au paramètre
      */
     public static double asDouble(int q28_4){
-        return Math.scalb(q28_4, -4);
+        return Math.scalb((double) q28_4, -4);
     }
 
     /**
